+++ workalendar/asia/malaysia.py
@@ -37,8 +37,8 @@ class Malaysia(ChineseNewYearCalendar, WesternCalendar, IslamicMixin):
         2015: date(2015, 11, 10),
         2016: date(2016, 10, 29),
         2017: date(2017, 10, 18),
-        2018: date(2018, 11, 7),   # This might change
-        2019: date(2019, 10, 27),  # This might change
+        2018: date(2018, 11, 6),
+        2019: date(2019, 10, 27),
         2020: date(2020, 11, 14),  # This might change
     }
 
@@ -51,9 +51,9 @@ class Malaysia(ChineseNewYearCalendar, WesternCalendar, IslamicMixin):
         2015: date(2015, 2, 3),
         2016: date(2016, 1, 25),
         2017: date(2017, 2, 9),
-        2018: date(2018, 1, 31),   # This might change
+        2018: date(2018, 1, 31),
         2019: date(2019, 1, 21),
-        2020: date(2020, 2, 8),
+        2020: date(2020, 2, 8),  # This might change
     }
     chinese_new_year_label = "First Day of Lunar New Year"
     include_chinese_second_day = True