+++ Changelog.md
@@ -5,6 +5,7 @@
 - Drop Python 3.4 support (#352).
 - Added Florida specific calendars: Florida Legal, Florida Circuit Courts, Miami-Dade (#216).
 - Added Malaysia Thaipusam days for the year 2019 & 2020 - thx @burlak for the bug report (#354).
+- Fixed Deepavali dates for the year 2018 ; confirmed fixed dates that were set in the past.
 
 ## v4.4.0 (2019-05-17)